Now uses Streams and Collections!
Changed the Brick[][] to an ArrayList<Brick> then used Java 8’s new
Stream API to do collision detection. This results in an overall more
responsive game, as it is doing the 15 brick tests in parallel, and
then only performing the collision action on the set of bricks that it
did collide with.